Reference

GinzburgLandau.GridFD

Base.reshape

GinzburgLandau.Laplacian2DPeriodic

GinzburgLandau.initial_conditions

GinzburgLandau.reaction

GinzburgLandau.GridFD — Type

GridFD

2D dicretization grid for the finite difference approach

Initializiation

GridFD(x::AbstractRange, y::AbstractRange)

Fields

x x coordinate, a $n$-long range

y y coordinate, a $n$-long range

h grid spacing

n number of grid points along one side

N total number of grid points

x_grid meshgrid for x, this is a $n imes n$ matrix

y_grid meshgrid for y

Base.reshape — Method

reshape(x::AbstractArray, g::AbstractGrid)

Reshapes between 2D array and a flattened 1D array.

GinzburgLandau.Laplacian2DPeriodic — Method

Laplacian2DPeriodic(g::GridFD)

Returns 2d finite diffence laplace operator for a domain of side length n with spacing dx as a sparse matrix. Periodic boundary condtions are applied

GinzburgLandau.initial_conditions — Method

initial_conditions(n::Integer)

Returns (random) initial conditions that are commonly used, 0.01*(rand(ComplexF64, (n,n)) .- (0.5 + 0.5im))

GinzburgLandau.reaction — Method

reaction(x, β)

The reaction term of the equation: $u - (1+ieta)|u|^2u$